// src/lib/contractorPlus/types/lead.d.ts
// Generated by https://quicktype.io
export interface LeadList {
	list: Lead[];
}
export interface Lead {
	lead_id?: string;
	salutation?: string;
	first_name?: string;
	last_name?: string;
	title?: string;
	source: string;
	description?: string;
	company?: string;
	website?: string;
	address?: string;
	address_2?: string;
	city?: string;
	state?: string;
	zip?: string;
	country?: string;
	tags?: string;
	assigned?: string;
	is_public?: string;
	phone_numbers?: PhoneNumber[];
	emails?: Email[];
}

export interface Email {
	value: string;
	is_primary: string;
}

export interface PhoneNumber {
	value: string;
	is_primary: string;
	code: string;
}